requirements.txt
Django==5.2.3
djangorestframework==3.16.0
dj-database-url==3.0.0
django-split-settings==1.3.2
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.32.4
idna==3.10
urllib3==2.4.0
certifi==2025.4.26

# Paiement
stripe==12.2.0
paypalrestsdk==1.13.3

# PDF et texte arabe
arabic-reshaper==3.0.0
python-bidi==0.6.6
pillow==11.3.0
reportlab==4.4.1
xhtml2pdf==0.2.17
svglib==1.5.1
tinycss2==1.4.0
lxml==5.4.0

# PDF signing (si utilisé)
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
oscrypto==1.3.0
cryptography==45.0.4
pyOpenSSL==25.1.0

# Date/heure
tzdata==2025.2
tzlocal==5.3.1
